rename files model to "*.model.ts", changes error display, uninstall "body-parser
