+++ package.json
@@ -29,7 +29,6 @@
     "@hapi/joi": "^17.1.1",
     "bcrypt": "^5.0.1",
     "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "debug": "^4.3.2",
