+++ src/index.ts
@@ -1,6 +1,5 @@
 import dotenv from 'dotenv';
 import express from 'express';
-import bodyParser from 'body-parser';
 import cookieParser from 'cookie-parser';
 import cors from 'cors';
 import errorMiddleware from './middleware/error.middleware';
@@ -16,15 +15,15 @@ const port = process.env.PORT || 8080;
 
 app.use(express.json());
 app.use(morgan('tiny'));
-app.use(express.static('public'));
+//app.use(express.static('public'));
 
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
+app.use(express.urlencoded({ extended: false }));
+app.use(express.json());
 app.use(cookieParser());
 
 app.use(
   cors({
-    origin: (origin, callback) => callback(null, true),
+    origin: (_origin, callback) => callback(null, true),
     credentials: true,
   }),
 );
