+++ src/validators/validate.middleware.ts
@@ -42,6 +42,7 @@ export const validateUserRegister: ValidationChain[] = [
 export const validateUserLogin: ValidationChain[] = [validateLogin, validatePassword];
 
 const validateNameRequired = body('name').isString().withMessage(messageString).notEmpty().withMessage(messageEmpty);
+
 const validateNameOptional = body('name')
   .isString()
   .withMessage(messageString)
@@ -57,18 +58,29 @@ const validatePhoto = body('photo')
   .withMessage(`Photo must be in the form of a URL`)
   .optional();
 
-const validateRecipe = body('recipe').isString().withMessage(messageString).optional();
+const validateRecipeRequired = body('recipe')
+  .isString()
+  .withMessage(messageString)
+  .notEmpty()
+  .withMessage(messageEmpty);
+
+const validateRecipeOptional = body('recipe')
+  .isString()
+  .withMessage(messageString)
+  .notEmpty()
+  .withMessage(messageEmpty)
+  .optional();
 
 export const validateCreateRecipe: ValidationChain[] = [
   validateNameRequired,
   validateType,
   validatePhoto,
-  validateRecipe,
+  validateRecipeRequired,
 ];
 
 export const validateUpdateRecipe: ValidationChain[] = [
   validateNameOptional,
   validateType,
   validatePhoto,
-  validateRecipe,
+  validateRecipeOptional,
 ];
